requirements.txt
mysqlclient==2.2.0
Django~=4.2
python-dotenv==1.0.0
requests==2.32.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
